refactor: move test-consent flow to utils
