+++ test/integration/integrationUtils.ts
@@ -1,6 +1,10 @@
 import got from 'got'
 
 const E2E_SIMULATOR_URL = 'http://localhost:1337'
+import { consents } from '../../lib/adapters/mydata'
+import { Login } from '../../lib/__generated__/myskills'
+import { defaultRequest } from '../../lib/services/consents'
+import { getConsentRequest } from '../../lib/services/db'
 
 export interface options {
   context: any
@@ -28,7 +32,25 @@ export const getClient = (
   return { mutate: client.mutate, query: client.query }
 }
 
-export const createMyDataAccount = (
+export const getConsentedClient = async (
+  server
+): Promise<{ query: Function; mutate: Function }> => {
+  await createMyDataAccount()
+  const request = defaultRequest(3600 * 24 * 31)
+  const { id } = await consents.request<Login>(request)
+  await approveConsent(id)
+  const { accessToken } = await getConsentRequest(id)
+
+  return getClient(server, {
+    context: {
+      headers: {
+        token: accessToken,
+      },
+    },
+  })
+}
+
+const createMyDataAccount = (
   firstName: string = 'Gordon',
   lastName: string = 'Freeman'
 ) => {
@@ -41,10 +63,10 @@ export const createMyDataAccount = (
         lastName,
       },
     },
-  }).then(data => data.body.id)
+  })
 }
 
-export const approveConsent = async consentRequestId => {
+const approveConsent = async consentRequestId => {
   const {
     body: { data },
   } = await got(`${E2E_SIMULATOR_URL}/getConsentRequest`, {
@@ -54,7 +76,8 @@ export const approveConsent = async consentRequestId => {
       args: consentRequestId,
     },
   })
-  const hej = await got(`${E2E_SIMULATOR_URL}/approveConsentRequest`, {
+
+  return got(`${E2E_SIMULATOR_URL}/approveConsentRequest`, {
     json: true,
     method: 'post',
     body: {
